// bindings/typescript/types/MikanMathTypes.ts
export class MikanVector2d {
  x: number = 0;
  y: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'x', type: 'double' },
    { name: 'y', type: 'double' }
  ];
}

export class MikanVector2f {
  x: number = 0;
  y: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'x', type: 'float' },
    { name: 'y', type: 'float' }
  ];
}

export class MikanVector2i {
  x: number = 0;
  y: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'x', type: 'int32' },
    { name: 'y', type: 'int32' }
  ];
}

export class MikanVector3d {
  x: number = 0;
  y: number = 0;
  z: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'x', type: 'double' },
    { name: 'y', type: 'double' },
    { name: 'z', type: 'double' }
  ];
}

export class MikanVector3f {
  x: number = 0;
  y: number = 0;
  z: number = 0;

  static __serializationMetadata: SerializationField[] = [
    { name: 'x', type: 'float' },
    { name: 'y', type: 'float' },
    { name: 'z', type: 'float' }
  ];
}